fix errors:
    = help: items from traits can only be used if the trait is in scope
help: the following trait is implemented but not in scope; perhaps add a `use` for it:
    |
1   | use substrate_stellar_sdk::xdr::impls::hash::ComputeHash;
