+++ src/xdr/impls/generalized_transaction_set.rs
@@ -1,27 +1,14 @@
 use crate::{
     compound_types::UnlimitedVarArray,
     types::{GeneralizedTransactionSet, TransactionPhase, TxSetComponent},
-    xdr::impls::hash::ComputeHash,
     Hash, TransactionEnvelope, XdrCodec,
 };
-use sha2::{Digest, Sha256};
-
-#[cfg(feature = "std")]
-impl ComputeHash for GeneralizedTransactionSet {
-    fn compute_hash(&self) -> Option<Hash> {
-        use std::convert::TryInto;
-
-        let mut hasher = Sha256::new();
-        hasher.update(self.to_xdr());
-
-        hasher.finalize().as_slice().try_into().ok()
-    }
-}
 
 impl GeneralizedTransactionSet {
     #[cfg(feature = "std")]
-    pub fn to_base64_encoded_xdr_string(&self) -> Result<&str, core::str::Utf8Error> {
-        std::str::from_utf8(&self.to_base64_xdr())
+    pub fn to_base64_encoded_xdr_string(&self) -> Result<String, std::string::FromUtf8Error> {
+        let base_64 = self.clone().to_base64_xdr();
+        String::from_utf8(base_64)
     }
 
     pub fn previous_ledger_hash(&self) -> Option<Hash> {
@@ -62,13 +49,13 @@ impl GeneralizedTransactionSet {
 
 #[cfg(test)]
 mod tests {
-    use crate::{types::GeneralizedTransactionSet, XdrCodec};
+    use crate::{types::GeneralizedTransactionSet, Hash, XdrCodec};
 
     const GENERALIZED_TX_SET: &str = "AAAAAVUcNbdctOS+OftZngTJY07YUUqb4P1I/owUmgdMuzscAAAAAgAAAAAAAAABAAAAAAAAAAEAAAAAAAAAZAAAAAEAAAACAAAAAMgOELhl9VFf5x0pG1aY8Mm/QQcnigdQ9MgWM1F8c6HSAAAAZAAZMt8AADrRAAAAAQAAAAAAAAAAAAAAAGUb2+8AAAABAAAAG3RzOjIwMjMtMTAtMDNUMDk6MTU6MDEuNTkyWgAAAAABAAAAAAAAAAEAAAAAf4MDV2AZH1oB1nouL9LSGUHGGafzcb48GXQyWFd9zswAAAAAAAAAAACYloAAAAAAAAAAAXxzodIAAABAEq3w/8HQ6kjqooVJPjg1TquL2pMOT+P9P7a3HpdqUYyFyJ8F32igbhIu3jvIJkafhDTosuL/rid2BxmScxhfDwAAAAAAAAAA";
 
-    const fn example() -> GeneralizedTransactionSet {
-        let to_vec_u8 = GENERALIZED_TX_SET.as_bytes();
-        GeneralizedTransactionSet::from_base64_xdr(to_vec_u8).expect("should be able to convert to txset")
+    fn example() -> GeneralizedTransactionSet {
+        let to_vec_u8 = GENERALIZED_TX_SET.as_bytes().to_vec();
+        GeneralizedTransactionSet::from_base64_xdr(to_vec_u8).expect("should return a GeneralizedTransactionSet")
     }
 
     #[test]
@@ -83,6 +70,7 @@ mod tests {
     #[test]
     fn generalized_transaction_set_returns_prev_ledger_hash() {
         let expected_hash = "VRw1t1y05L45+1meBMljTthRSpvg/Uj+jBSaB0y7Oxw=".as_bytes();
+        let expected_hash = Hash::from_base64_xdr(expected_hash).expect("return hash");
 
         let gen_tx_set = example();
         let actual_hash = gen_tx_set.previous_ledger_hash().expect("should return a previous ledger hash");
