+++ src/xdr/impls/hash.rs
@@ -1,5 +1,8 @@
 use crate::{AsBinary, Hash, StellarSdkError};
 
+#[cfg(all(feature = "all-types", feature = "std"))]
+use crate::types::{GeneralizedTransactionSet, TransactionSet};
+
 pub trait IntoHash {
     fn into_hash(self) -> Result<Hash, StellarSdkError>;
 }
@@ -19,4 +22,36 @@ impl<T: AsRef<[u8]>> IntoHash for AsBinary<T> {
 /// Returns the hash of the Stellar object
 pub trait ComputeHash {
     fn compute_hash(&self) -> Option<Hash>;
+}
+
+#[cfg(all(feature = "all-types", feature = "std"))]
+impl ComputeHash for GeneralizedTransactionSet {
+    fn compute_hash(&self) -> Option<Hash> {
+        use crate::XdrCodec;
+        use sha2::{Digest, Sha256};
+        use std::convert::TryInto;
+
+        let mut hasher = Sha256::new();
+        hasher.update(self.to_xdr());
+
+        hasher.finalize().as_slice().try_into().ok()
+    }
+}
+
+#[cfg(all(feature = "all-types", feature = "std"))]
+impl ComputeHash for TransactionSet {
+    fn compute_hash(&self) -> Option<Hash> {
+        use crate::XdrCodec;
+        use sha2::{Digest, Sha256};
+        use std::convert::TryInto;
+
+        let mut hasher = Sha256::new();
+        hasher.update(self.previous_ledger_hash);
+
+        self.txes.get_vec().iter().for_each(|envlp| {
+            hasher.update(envlp.to_xdr());
+        });
+
+        hasher.finalize().as_slice().try_into().ok()
+    }
 }
